fix(types-1): Spark-door Arrow types follow Spark — INT literals, count/sum widths, rank family, from_unixtime (#386)
* docs(types-1): charter ledger and staging map row

pins: types-1/C-001, C-002, C-003, C-004, C-005, C-006, C-007, C-008, C-009
Authored-By: Muse Spark (muse-spark-1.3) <noreply@meta.ai>

* test(types-1): red-first Spark-door Arrow-type pins

25 red on the base (literals, count-likes, rank, from_unixtime, casts);
28 green controls pin what already matches Spark.
pins: types-1/C-001, C-002, C-003, C-004, C-005, C-006, C-007, C-008
Authored-By: Muse Spark (muse-spark-1.3) <noreply@meta.ai>

* feat(types-1): Spark-door Arrow types follow Spark

Integer literals narrow to INT, count-likes answer BIGINT, the rank
family answers INT, from_unixtime answers session-zone STRING. The
narrowing rule runs after TypeCoercion with a closing coercion pass so
re-opened mixes never reach physical planning; LIMIT fetch/skip stay
Int64; plain INSERT conforms narrowed INT sources into BIGINT targets.
pins: types-1/C-001, C-002, C-003, C-004, C-005, C-006, C-007
Authored-By: Muse Spark (muse-spark-1.3) <noreply@meta.ai>

* test(types-1): triage facade suite to Spark answers

76 failures classified flip-or-revert: lawful flips pin the Spark
answer with citation; stale Spark-2.x-era no-engine tiers removed
from window parity; null-top last() corrected on both engines.
check_lib_py baselines amended (core.py +2, window parity ratchet).
pins: types-1/C-008
Authored-By: Muse Spark (muse-spark-1.3) <noreply@meta.ai>

* test(types-1): mutation score 7/7 bite, 0 survivors

M1-M5 at Rust level, M6 end-to-end wiring via test_types_1.py.
pins: types-1/C-008
Authored-By: Muse Spark (muse-spark-1.3) <noreply@meta.ai>

* fix(types-1): from_unixtime nullable on both doors, like Spark

Live oracle (4.1.2) marks FromUnixTime nullable even for non-null
input; the new UDF derived non-null and the live leg reddened.
pins: types-1/C-006
Authored-By: Muse Spark (muse-spark-1.3) <noreply@meta.ai>

* test(types-1): full live-oracle cover for C-001..C-006, residues TY-6/BL-18

Four live legs on PySpark 4.1.2/UTC: full-match shapes across
literals, arithmetic, aggregates, rank and from_unixtime; approx/regr
nullability carved out and pinned; ANSI-off overflow leg. Residues
filed as TY-6 (UNION width) and BL-18 (approx/regr nullability).
pins: types-1/C-009
Authored-By: Muse Spark (muse-spark-1.3) <noreply@meta.ai>

* docs(types-1): close the unit — 9 PROVEN, ledger to completed

Verdicts with measured cells, §9 live-oracle table (Spark 4.1.2/UTC),
§12 coverage attestation, §6 true-up. Staging row to completed.
pins: types-1/C-009
Authored-By: Muse Spark (muse-spark-1.3) <noreply@meta.ai>

* docs(types-1): sweep eight comment leaks into map.md rows

The round-3 hand-back carried eight new comment lines in code, which the
brief forbids. The lib.rs analyzer_rules doc restores origin/main's text
verbatim; the other seven lines are deleted, with each reason relocated
to its directory map row (rule order, BL-8/UNIX-1 retirements, TY-3
nullability residue, CTAS Int32) or an assertion message.
pins: types-1/C-002, C-003, C-006, C-007
Authored-By: Muse Spark (muse-spark-1.3) <noreply@meta.ai>

* fix(types-1): from_unixtime wraps like Spark; '+' for 5-digit years

Live PySpark 4.1.2 wraps out-of-range seconds (i64::MAX renders -1s),
exactly like the shipped wrapping_mul — saturation was never the gap.
The real gap was Java's '+' sign on 5+-digit years (+51190-09-21):
the shared Java-pattern year arm now emits it for count >= 4 when
digits exceed the width. Rust wrap test; datetime.rs 1704->1709
(INCREASE, owner approval at merge).
pins: types-1/C-006
Authored-By: Muse Spark (muse-spark-1.3) <noreply@meta.ai>

* test(types-1): round-4 pins — extremes, re-coercion, grouping, sum, ntile, formats

Spark-measured cells on both doors for every critic shape: wrapped-year
from_unixtime extremes, CASE/COALESCE/IF/array/struct/map/UNION/DECIMAL
re-coercion, ROLLUP/SETS grouping with the (int32, int8) carve-out,
TINYINT/SMALLINT sums, ntile(BIGINT) acceptance, EEE and dd/MM formats.
Three live legs added (seven total); TY-7/8/9/10 filed. Also triages
three main-merge avg pins whose VALUES k narrowed to int32 (lawful,
C-001's live legs are the citation).
pins: types-1/C-001, C-002, C-004, C-005, C-006
Authored-By: Muse Spark (muse-spark-1.3) <noreply@meta.ai>

* fix(types-1): with_row_index stays BIGINT

row_number() narrowed to Int32 silently shifted the polars index column
int64 -> int32; cast to BIGINT like the core.py sample paths and pin
the index type with offset values.
pins: types-1/C-005
Authored-By: Muse Spark (muse-spark-1.3) <noreply@meta.ai>

* docs(types-1): round-4 corrections — ledger, citations, ceiling, maps

F2/F3: section 11 rewritten to the verified triage diff (eight
converged tiers, no last() change anywhere; core.py +2 is the two
__repark_rn BIGINT casts); C-004/C-007 and section 7.1 restated to the
built mechanism (AnalyzerRule narrowing, UDF-registration wrappers);
BL-8 and G5-RANK-TYPE-1 restated likewise; new section 13 records all
eleven findings with the round-4 oracle table. F6: three inverted pins
repointed to types-1/C-002 with the F-Y10-1 supersession noted. F10:
one import joined absorbs core.py 6305->6303 in both cap tables.
F9/F11 verified no-change. Maps in lockstep.
pins: types-1/C-001, C-002, C-003, C-004, C-005, C-006, C-007, C-008, C-009
Authored-By: Muse Spark (muse-spark-1.3) <noreply@meta.ai>

* test(types-1): the dbt cursor description reads a VALUES literal as int32, Spark's answer

* fix(types-1): negative years pad digits after the sign; arm moves to spark_year_pad

The Java-pattern year arm pads the digits and re-attaches the sign
(-0499, -0002), yy is abs(year) % 100 (-499 goes 01 to 99), and 5+
digit positives keep the plus. datetime.rs 1709 to 1700 via the
extraction; both cap tables mirrored in this commit. Every cell
live-measured on Spark 4.1.2, ANSI on and off, UTC and New York.
pins: types-1/C-006
Authored-By: Muse Spark (muse-spark-1.3) <noreply@meta.ai>

* test(types-1): negative 3/4-digit-year from_unixtime cells on both doors

Default, yyyy and yy spellings for -77900000000 (-0499) and -62200000000
(-0002) pinned on the SQL door and the facade in UTC, one New York cell
on both doors, and the extremes live leg extended over the UTC cells.
Spark 4.1.2 answers, ANSI on and off.
pins: types-1/C-006
Authored-By: Muse Spark (muse-spark-1.3) <noreply@meta.ai>

* docs(types-1): round-5 ledger - N2/N3 corrections, dbt gate, oracle cells

N2: section 6 and the dataframe map now name the real core.py change
(the __repark_rn BIGINT casts), as section 11 already reads. N3: C-008
carries make py-test-dbt in its roster and evidence, naming the cursor
description int64 to int32 flip (Spark's answer per C-001, ffae551d).
New section 14 records the four dispositions with the oracle table.
pins: types-1/C-001, C-006, C-008
Authored-By: Muse Spark (muse-spark-1.3) <noreply@meta.ai>

* docs(types-1): §13.2 points forward to the 1700 ceiling

+++ python/repark/tests/test_sql_passthrough_parity.py
@@ -115,26 +115,26 @@ def test_decimal_division_stays_decimal(spark: ReparkSession) -> None:
 
 def test_order_by_asc_defaults_nulls_first(spark: ReparkSession) -> None:
     result = spark.sql("SELECT v FROM (VALUES (2), (NULL), (1)) AS t(v) ORDER BY v")
-    golden = pa.table({"v": pa.array([None, 1, 2], pa.int64())})
+    golden = pa.table({"v": pa.array([None, 1, 2], pa.int32())})
     assert_frames_equal(result.to_arrow(), golden, order_sensitive=True)
 
 
 def test_order_by_desc_defaults_nulls_last(spark: ReparkSession) -> None:
     result = spark.sql("SELECT v FROM (VALUES (2), (NULL), (1)) AS t(v) ORDER BY v DESC")
-    golden = pa.table({"v": pa.array([2, 1, None], pa.int64())})
+    golden = pa.table({"v": pa.array([2, 1, None], pa.int32())})
     assert_frames_equal(result.to_arrow(), golden, order_sensitive=True)
 
 
 def test_order_by_explicit_null_placement_is_honoured(spark: ReparkSession) -> None:
     result = spark.sql("SELECT v FROM (VALUES (2), (NULL), (1)) AS t(v) ORDER BY v ASC NULLS LAST")
-    golden = pa.table({"v": pa.array([1, 2, None], pa.int64())})
+    golden = pa.table({"v": pa.array([1, 2, None], pa.int32())})
     assert_frames_equal(result.to_arrow(), golden, order_sensitive=True)
 
 
 def test_order_by_limit_returns_spark_rows(spark: ReparkSession) -> None:
     # The data-changing case: under LIMIT the placement decides WHICH rows survive.
     result = spark.sql("SELECT v FROM (VALUES (2), (NULL), (1)) AS t(v) ORDER BY v LIMIT 2")
-    golden = pa.table({"v": pa.array([None, 1], pa.int64())})
+    golden = pa.table({"v": pa.array([None, 1], pa.int32())})
     assert_frames_equal(result.to_arrow(), golden, order_sensitive=True)
